Refactor Itinerary Section into JS Module Structure
